chore: Use the `AppConfigurationOffline` implementation in testing
Tests use a hardcoded JSON file to read the configuration from it. The offline
implementation provides exactly this functionality.
